[052_modular_approach], issue #52, Moved classes for export data to dedicated module
To improve code structure, the classes 'EdgeData', 'FaceData' and 'BoundaryData', as well as the functions that were included in the 'geom_extractor.py' module, have been moved into the dedicated 'export_data.py' module.
